example: ipfragreass: avoid undefined behavior in bit shift in hash()
Avoid undefined behavior in bit shift in hash() by casting to an
unsigned integer of sufficient width before the shift operation.

Fixes GCC undefined sanitizer error:

odp_ipfragreass_reassemble.c:92:23: runtime error: left shift of 40448 by 16 places cannot be represented in type 'int'

+++ example/ipfragreass/odp_ipfragreass_reassemble.c
@@ -89,7 +89,7 @@ static inline uint32_t hash(odph_ipv4hdr_t *hdr)
 {
 	uint32_t a = hdr->src_addr;
 	uint32_t b = hdr->dst_addr;
-	uint32_t c = hdr->id << 16 | hdr->proto;
+	uint32_t c = (uint32_t)hdr->id << 16 | hdr->proto;
 
 	/* A degenerate 3x32-bit Jenkins hash */
 	c ^= b;
